+++ app/vite.config.mjs
@@ -39,7 +39,7 @@ export default defineConfig(async () => ({
 
   clearScreen: false,
   server: {
-    port: 1420,
+    port: 1425,
     strictPort: true,
     host: host || false,
     hmr: host